+++ CHANGELOG.md
@@ -224,8 +224,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [SIL.NUnit3Compatibility] new project/package that allows to use NUnit3 syntax with NUnit2
   projects
 
-[Unreleased]: https://github.com/sillsdev/libpalaso/compare/v10.0.0...master
+[Unreleased]: https://github.com/sillsdev/libpalaso/compare/v10.1.0...master
 
+[10.1.0]: https://github.com/sillsdev/libpalaso/compare/v10.0.0...v10.1.0
 [10.0.0]: https://github.com/sillsdev/libpalaso/compare/v9.0.0...v10.0.0
 [9.0.0]: https://github.com/sillsdev/libpalaso/compare/v8.0.0...v9.0.0
 [8.0.0]: https://github.com/sillsdev/libpalaso/compare/v7.0.0...v8.0.0